+++ mirrord/layer/src/lib.rs
@@ -210,6 +210,13 @@ pub(crate) static ENABLED_UDP_OUTGOING: OnceLock<bool> = OnceLock::new();
 /// When true, localhost connections will stay local (won't go to the remote pod localhost)
 pub(crate) static OUTGOING_IGNORE_LOCALHOST: OnceLock<bool> = OnceLock::new();
 
+/// Tells us if the user enabled wants to ignore listening on localhost in [`IncomingConfig`].
+///
+/// ## Usage
+///
+/// When true, localhost connections will stay local (won't go to the remote pod localhost)
+pub(crate) static INCOMING_IGNORE_LOCALHOST: OnceLock<bool> = OnceLock::new();
+
 /// Check if we're running in NixOS or Devbox.
 ///
 /// - If so, add `sh` to the skip list because of
@@ -390,6 +397,10 @@ fn layer_start(config: LayerConfig) {
     OUTGOING_IGNORE_LOCALHOST
         .set(config.feature.network.outgoing.ignore_localhost)
         .expect("Setting OUTGOING_IGNORE_LOCALHOST singleton");
+
+    INCOMING_IGNORE_LOCALHOST
+        .set(config.feature.network.incoming.ignore_localhost)
+        .expect("Setting INCOMING_IGNORE_LOCALHOST singleton");
 
     FILE_FILTER.get_or_init(|| FileFilter::new(config.feature.fs.clone()));