+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## 3.31.0
+
 ### Added
 
 - config: `ignore_localhost` to `outgoing` config for ignoring localhost connections, meaning it will connect to local